Split format_sql tests
