+++ test/unit/Jamfile
@@ -84,6 +84,12 @@ run
         test/spotchecks/multifn.cpp
         test/spotchecks/misc.cpp
 
+        test/format_sql/formattable.cpp
+        test/format_sql/basic_format_context.cpp
+        test/format_sql/individual_value.cpp
+        test/format_sql/format_strings.cpp
+        test/format_sql/alternate_apis.cpp
+
         test/execution_state.cpp
         test/static_execution_state.cpp
         test/results.cpp
@@ -114,7 +120,6 @@ run
         test/character_set.cpp
         test/escape_string.cpp
         test/constant_string_view.cpp
-        test/format_sql.cpp
         
     : requirements
         <toolset>msvc:<cxxflags>-FI"pch.hpp" # https://github.com/boostorg/boost/issues/711
